fix #8

+++ datasets/data_augment.py
@@ -272,6 +272,6 @@ class preproc(object):
         landm_t[:, 0::2] /= width
         landm_t[:, 1::2] /= height
         labels_t = np.expand_dims(labels_t, 1)
-        targets_t = np.hstack((boxes_t, landm_t, labels_t))
+        targets_t = np.hstack((boxes_t, landm_t, labels_t)).astype(np.float32)
 
         return image_t, targets_t
